Add correct line parsing

+++ pandora.py
@@ -9,33 +9,29 @@ import threading
 from signal import SIGTERM
 from input_classes import Task, Expression
 
-
-def getPlayList():
-    titleRe = compile(r"\s+0\)\s+.+")
+def getStation(title):
+    titleRe = compile(r"\s+(\d+)\)\s+.+")
     pandora = subprocess.Popen(
         ["pianobar"], stdout=subprocess.PIPE,  shell=True, preexec_fn=os.setsid)
     while True:
-        line = pandora.stdout.readline().decode('utf-8')
-        if titleRe.match(line):
-            yield line.strip()
-        elif line.startswith("[?]"):
+        # remove ANSI sequence before and newline after
+        line = pandora.stdout.readline().decode('utf-8')[6:-1]
+        m = titleRe.match(line)
+        if title.lower() in line.lower() and m:
+            return m.group(1)
+        if line.startswith("[?]"):
             os.killpg(os.getpgid(pandora.pid), SIGTERM)
-
+    return ""
 
 def pandorify(args, chan):
     """ Return the first youtube result for a link """
     print("Started", args['station'])
     num_extractor = compile(r"(\d+)\).+")
-    track_num = ""
-    for title in getPlayList():
-        if args['station'] in title:
-            track_num = num_extractor.match(title).group(1)
+    track_num = getStation(args['station'])
     if track_num == "":
         print("Bad search pattern")
         return
-    with open(os.devnull, 'wb') as throw_away:
-        pandora = subprocess.Popen(
-            ["pianobar"], stdin=subprocess.PIPE,  shell=True, preexec_fn=os.setsid, stdout=throw_away)
+    pandora = subprocess.Popen(["pianobar"], stdin=subprocess.PIPE)
     pandora.stdin.write(bytes(track_num+"\n", "UTF-8"))
     command_dict = {
         "pause": "p",
@@ -59,15 +55,16 @@ def pandorify(args, chan):
             return
 
 PANDORA_TASK = Task("youtube",
-               [Expression(compile(r"play (.+) radio on pandora"), ('station',)),
-                Expression(compile(r"play (.+) on pandora"), ('station',)),
-                ],
-               [Expression(compile("pause"), ("command",)),
-                Expression(compile("play"), ("command",)),
-                Expression(compile("like"), ("command",)),
-                Expression(compile("thumbs up"), ("command",)),
-                Expression(compile("unlike"), ("command",)),
-                Expression(compile("thumbs down"), ("command",)),
-                Expression(compile("next"), ("command",))
-                ],
-               pandorify)
+                    [Expression(compile(r"play (.+) radio on pandora"), ('station',)),
+                     Expression(compile(r"play (.+) on pandora"),
+                                ('station',)),
+                     ],
+                    [Expression(compile("pause"), ("command",)),
+                        Expression(compile("play"), ("command",)),
+                        Expression(compile("like"), ("command",)),
+                        Expression(compile("thumbs up"), ("command",)),
+                        Expression(compile("unlike"), ("command",)),
+                        Expression(compile("thumbs down"), ("command",)),
+                        Expression(compile("next"), ("command",))
+                     ],
+                    pandorify)
